% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\name{cmat}
\alias{cmat}
\alias{cmat<-}
\alias{contr.hypothesis}
\title{Retrieve or set contrast matrix}
\usage{
cmat(x, add_intercept = FALSE, remove_intercept = FALSE)

cmat(x, add_intercept = FALSE, remove_intercept = FALSE) <- value

contr.hypothesis(..., add_intercept = FALSE, remove_intercept = FALSE)
}
\arguments{
\item{x}{A hypr object}

\item{value}{contrast matrix}
}
\description{
Use these functions to retrieve or set a hypr object's contrast matrix. If used for updating, the hypothesis matrix and equations are derived automatically.
}
\section{Functions}{
\itemize{
\item \code{cmat<-}: Set contrast matrix

\item \code{contr.hypothesis}: Retrieve contrast matrix to override factor contrasts
}}

