% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\name{conversions}
\alias{conversions}
\alias{eqs2hmat}
\alias{eqs2cmat}
\alias{hmat2cmat}
\alias{cmat2hmat}
\alias{hmat2eqs}
\alias{cmat2eqs}
\title{Shorthand versions for simple hypothesis translation}
\usage{
eqs2hmat(eqs, terms = NULL)

eqs2cmat(eqs)

hmat2cmat(hmat)

cmat2hmat(cmat)

hmat2eqs(hmat, as_fractions = TRUE)

cmat2eqs(cmat)
}
\arguments{
\item{eqs}{A list() of equations}

\item{terms}{(optional) A character vector of variables to be expected (if not provided, automatically generated from all terms occurring in the equations list)}

\item{hmat}{A hypothesis matrix}

\item{cmat}{A contrast matrix}
}
\value{
A list of equations (hmat2eqs and cmat2eqs), a contrast matrix (hmat2cmat, eqs2cmat), or a hypothesis matrix (cmat2hmat, eqs2hmat)
}
\description{
Shorthand versions for simple hypothesis translation
}
\section{Functions}{
\itemize{
\item \code{eqs2hmat}: Convert null hypothesis equations to hypothesis matrix

\item \code{eqs2cmat}: Convert null hypothesis equations to contrast matrix

\item \code{hmat2cmat}: Convert hypothesis matrix to contrast matrix

\item \code{cmat2hmat}: Convert contrast matrix to hypothesis matrix

\item \code{hmat2eqs}: Convert hypothesis matrix to null hypothesis equations

\item \code{cmat2eqs}: Convert contrast matrix to null hypothesis equations
}}

