% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\name{formula<-}
\alias{formula<-}
\title{Manipulate the formulas of an S4 object}
\usage{
formula(x, ...) <- value
}
\arguments{
\item{x}{The object to manipulate}

\item{...}{Additional arguments passed on to the method}

\item{value}{The new formula}
}
\description{
This is a generic function for setting an S4 object’s formulas.
}
