% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\name{ginv2}
\alias{ginv2}
\title{Nice generalized inverse}
\usage{
ginv2(x, as_fractions = TRUE)
}
\arguments{
\item{x}{The original matrix}

\item{as_fractions}{Whether to format the matrix as fractions (MASS package)}
}
\value{
Nice generalized inverse
}
\description{
This function calculates the generalized inverse of x, formats it as fractions and copies dimension names from the original matrix.
}
\seealso{
MASS::ginv()
}
