% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\docType{class}
\name{hypr-class}
\alias{hypr-class}
\alias{show,hypr-method}
\title{An object containing null hypotheses, a hypothesis matrix, and a contrast matrix}
\usage{
\S4method{show}{hypr}(object)
}
\description{
An object containing null hypotheses, a hypothesis matrix, and a contrast matrix
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print summary of hypr object
}}

\section{Slots}{

\describe{
\item{\code{eqs}}{A list of null hypotheses}

\item{\code{hmat}}{A hypothesis matrix}

\item{\code{cmat}}{A contrast matrix}
}}

