% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\docType{methods}
\name{terms,hypr-method}
\alias{terms,hypr-method}
\title{Retrieve the terms (variables) used in a hypr object}
\usage{
\S4method{terms}{hypr}(x)
}
\arguments{
\item{x}{A hypr object}
}
\value{
A character vector of term names
}
\description{
Retrieve the terms (variables) used in a hypr object
}
