% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\name{hmat}
\alias{hmat}
\alias{thmat}
\alias{hmat<-}
\alias{thmat<-}
\title{Retrieve and set hypothesis matrix}
\usage{
hmat(x, as_fractions = TRUE)

thmat(x, as_fractions = TRUE)

hmat(x) <- value

thmat(x) <- value
}
\arguments{
\item{x}{A hypr object}

\item{as_fractions}{Whether to format matrix as fractions (MASS package)}

\item{value}{Hypothesis matrix}
}
\description{
Use these functions to retrieve or set a hypr object's hypothesis matrix. If used for updating, the contrast matrix and equations are derived automatically.
}
\section{Functions}{
\itemize{
\item \code{thmat}: Retrieve transposed hypothesis matrix

\item \code{hmat<-}: Set hypothesis matrix

\item \code{thmat<-}: Set transposed hypothesis matrix
}}

