% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypr.R
\name{hmat}
\alias{hmat}
\alias{hmat<-}
\title{Retrieve or set hypothesis matrix}
\usage{
hmat(x)

hmat(x) <- value
}
\arguments{
\item{x}{A hypr object}

\item{value}{Hypothesis matrix}
}
\description{
Description
}
\section{Functions}{
\itemize{
\item \code{hmat<-}: Set hypothesis matrix
}}

